NodeManager cleanup
